cleaned up repititive code in bayes and msm
