+++ msmbuilder/msm/bayesmsm.py
@@ -40,10 +40,8 @@ import itertools
 import warnings
 
 import numpy as np
-from ..utils import list_of_1d
 from ..base import BaseEstimator
-from .core import (_MappingTransformMixin, _dict_compose,
-                   _strongly_connected_subgraph, _transition_counts,
+from .core import (_MappingTransformMixin,
                    _solve_msm_eigensystem)
 from ._metzner_mcmc_fast import metzner_mcmc_fast
 from ._metzner_mcmc_slow import metzner_mcmc_slow
@@ -173,25 +171,7 @@ class BayesianMarkovStateModel(BaseEstimator, _MappingTransformMixin):
         self.percent_retained_ = None
 
     def fit(self, sequences, y=None):
-        sequences = list_of_1d(sequences)
-        if int(self.lag_time) <= 0:
-            raise ValueError('Invalid lag_time: %s' % self.lag_time)
-        raw_counts, mapping = _transition_counts(
-            sequences, int(self.lag_time), sliding_window=self.sliding_window)
-
-        ergodic_cutoff = self._parse_ergodic_cutoff()
-
-        if ergodic_cutoff > 0:
-            self.countsmat_, mapping2, self.percent_retained_ = \
-                _strongly_connected_subgraph(raw_counts,
-                                             ergodic_cutoff, self.verbose)
-            self.mapping_ = _dict_compose(mapping, mapping2)
-        else:
-            self.countsmat_ = raw_counts
-            self.mapping_ = mapping
-            self.percent_retained_ = 100
-
-        self.n_states_ = self.countsmat_.shape[0]
+        self._setup(sequences)
         fit_method_map = {
             True: self._fit_reversible,
             False: self._fit_non_reversible}
